+++ src/components/auth/signupModal.js
@@ -1,0 +1,66 @@
+import Swal from "sweetalert2";
+
+const signUpModal = () => {
+  Swal.fire({
+    title: 'Sign Up',
+    html: ` <input id="swal-input1" class="swal2-input" placeholder="First Name" required>
+            <input id="swal-input2" class="swal2-input" placeholder="Last Name" required>
+            <input id="swal-input3" class="swal2-input" placeholder="Email" required>
+            <input id="swal-input4" class="swal2-input" placeholder="Password" required>`,
+    focusConfirm: false,
+    showCancelButton: true,
+    confirmButtonText: 'Sign Up',
+    reverseButtons: true,
+    confirmButtonColor: '#99eb1b',
+    preConfirm: () => {
+      // validate required fields
+      if (!document.getElementById('swal-input1').value || !document.getElementById('swal-input2').value || !document.getElementById('swal-input3').value || !document.getElementById('swal-input4').value) {
+        Swal.showValidationMessage(`Please fill in all fields`)
+      }else{
+        // send data to backend
+        const firstName = document.getElementById('swal-input1').value;
+        const lastName = document.getElementById('swal-input2').value;
+        const email = document.getElementById('swal-input3').value;
+        const password = document.getElementById('swal-input4').value;
+        const data = {firstName, lastName, email, password}
+        // attempt signup
+        fetch('http://localhost:5000/api/v1/users/signup', {
+          method: 'POST',
+          headers: {
+            'Content-Type': 'application/json'
+          },
+          body: JSON.stringify(data)
+        })
+        .then(res => res.json())
+        .then(data => {
+          if(data.status === 'success'){
+            Swal.fire({
+              icon: 'success',
+              title: 'Success',
+              text: data.message,
+              confirmButtonColor: '#99eb1b',
+            })
+          }else{
+            Swal.fire({
+              icon: 'error',
+              title: 'Error',
+              text: data.message,
+              confirmButtonColor: '#99eb1b',
+            })
+          }
+        })
+        .catch(err => {
+          console.log(err);
+          Swal.fire({
+            icon: 'error',
+            title: 'Error',
+            text: 'Something went wrong. Please try again later',
+            confirmButtonColor: '#99eb1b',
+          });
+        })
+      }
+    },
+    allowOutsideClick: () => !Swal.isLoading()
+  });
+};
+export default signUpModal;
